Tests/LibWeb/Text/expected/Wasm/WebAssembly-grow-hook.txt

Size before grow: 65536
Size after grow, before refresh: 0
Size after grow, after refresh: 131072
